requirements.txt
# pip install -r requirements.txt

# Google API & Auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-genai

# Image Processing
Pillow

# Environment Variables
python-dotenv